requirements.txt
anaconda-anon-usage
archspec 
bidict==0.23.1
blinker==1.8.2
boltons 
Brotli 
certifi 
cffi 
charset-normalizer 
click==8.1.7
colorama 
conda 
conda-content-trust 
conda-libmamba-solver 
conda-package-handling 
conda_package_streaming 
cryptography
distro 
dnspython==2.6.1
eventlet==0.36.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
frozendict 
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
idna 
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch 
jsonpointer==2.1
libmambapy 
MarkupSafe==2.1.5
menuinst 
packaging 
platformdirs 
pluggy 
pycosat 
pycparser 
PySocks 
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
requests 
ruamel.yaml 
setuptools==69.5.1
simple-websocket==1.0.0
tqdm 
truststore 
urllib3 
waitress==3.0.0
Werkzeug==3.0.3
wheel==0.43.0
win-inet-pton 
wsproto==1.2.0
zope.event==5.0
zope.interface==7.0.1
zstandard 
